Checking if the file exists before deleting
Checking if the cached HTML snippet file exists, before deleting it during a new CSV upload.

+++ OneDash/AdminArea/Products.aspx.cs
@@ -159,7 +159,10 @@ public partial class AdminArea_Products : System.Web.UI.Page
                     var csvFilePath = Path.Combine(dirPath, "BugList.csv");
                     FURBugList.SaveAs(csvFilePath);
                     var cachedHTMLSnippetFilePath = Path.Combine(dirPath, "BugList.htmlSnippet");
-                    File.Delete(cachedHTMLSnippetFilePath);
+                    if (File.Exists(cachedHTMLSnippetFilePath))
+                    {
+                        File.Delete(cachedHTMLSnippetFilePath);
+                    }
                 }
                 else
                 {
